pyproject.toml
[project]
name = "pmer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.2",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-sqlalchemy>=3.1.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "pyotp>=2.9.0",
    "qrcode>=8.0",
    "pillow>=11.0.0",
]
